% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCarbonateEq.R
\name{CCarbonateEq}
\alias{CCarbonateEq}
\title{Carbonate equilibrium calculator}
\description{
Provides tools for calculating activities of inorganic carbon
  species based on equilibrium constants.

  Implementation is provided by C++ objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{externalPointer}}{The external pointer to the associated C++ CarbonateEq object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CCarbonateEq$new()}}
\item \href{#method-reset}{\code{CCarbonateEq$reset()}}
\item \href{#method-optfCO2FromDICTotalAlk}{\code{CCarbonateEq$optfCO2FromDICTotalAlk()}}
\item \href{#method-clone}{\code{CCarbonateEq$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCarbonateEq$new(tempC, eConduct = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tempC}}{A numerical value for the temperature}

\item{\code{eConduct}}{An optonal numerical value for the electrical conductivity
Default value is 0 (results in an ionic strength of 0)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the calculator to a new temperature and electrical conductivity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCarbonateEq$reset(tempC, eConduct = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tempC}}{A numerical value for the temperature}

\item{\code{eConduct}}{An optonal numerical value for the electrical conductivity
Default value is 0 (results in an ionic strength of 0)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optfCO2FromDICTotalAlk"></a>}}
\if{latex}{\out{\hypertarget{method-optfCO2FromDICTotalAlk}{}}}
\subsection{Method \code{optfCO2FromDICTotalAlk()}}{
Performs an optimization calculation to estimate the partial
  pressure of carbon dioxide based on a known DIC and total
  alkalinity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCarbonateEq$optfCO2FromDICTotalAlk(
  dic,
  totalAlk,
  pHTol = 1e-05,
  pHMin = 2,
  pHMax = 12
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dic}}{Concentration of DIC.
Units of molarity.}

\item{\code{totalAlk}}{Concentration of total alkalinity.
Units of molarity.}

\item{\code{pHTol}}{Optional value for tolerance of pH optimization.
Default value is 1e-5.}

\item{\code{pHMin}}{Optional value for the minimum pH value for optimization.
Default value is 2.}

\item{\code{pHMax}}{Optional value for the maximum pH value for optimization.
Default value is 12.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCarbonateEq$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
