% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetab.R
\name{CMetab}
\alias{CMetab}
\title{C++ model of aqueous metabolism influence on oxygen}
\description{
Abstract class (should not be instantiated on its own) defining
  the base functionality of a metabolism model in aquatic systems.

  Implementations of models derived from this class are provided
  by C++ objects. See \link{CMetabDo} as an example of a fully
  functional implementation of this abstraction.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointers}}{A list of external pointers to the C++ objects associated with
implementing the features of the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetab-run}{\code{CMetab$run()}}
\item \href{#method-CMetab-setMetabParam}{\code{CMetab$setMetabParam()}}
\item \href{#method-CMetab-clone}{\code{CMetab$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetab-run"></a>}}
\if{latex}{\out{\hypertarget{method-CMetab-run}{}}}
\subsection{Method \code{run()}}{
Runs a simulation based on parameters and driving data
  in the object attributes. Output will be stored in the
  output attribute, but is also returned from this function
  as a convenience.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetab$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The S expression object returned by the call to the C++ method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetab-setMetabParam"></a>}}
\if{latex}{\out{\hypertarget{method-CMetab-setMetabParam}{}}}
\subsection{Method \code{setMetabParam()}}{
Sets one of the model parameters to a new value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetab$setMetabParam(param, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A character string to indicate the parameter to be changed
\itemize{
  \item "DailyGPP": change the daily GPP parameter
  \item "DailyER": change the daily ER parameter
  \item "k600": change the k600 parameter
}}

\item{\code{value}}{The new value for the parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The previous value of the parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetab-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetab-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
