% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabDoDic.R
\name{CMetabDoDic}
\alias{CMetabDoDic}
\title{Time series DO and DIC metabolism model based on C++ implementation}
\description{
Predicts the change in dissolved oxygen and dissolved inorganic carbon
  in an aquatic system over time due to metabolic processes and
  gas exchange with the atmosphere.

  Implementation of this model is provided by C++ objects.
}
\section{Super class}{
\code{\link[metabc:CMetab]{metabc::CMetab}} -> \code{CMetabDoDic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "ForwardEuler" or "CrankNicolson" are known solutions)}

\item{\code{time}}{A numeric vector of time in units of days representing the
times at which DO concentrations are simulated}

\item{\code{timePOSIX}}{A vector of POSIXct type time objects representing the
times at which DO concentrations are simulated}

\item{\code{airPressure}}{A vector of barometric pressures at the times when DO
values are to be simulated}

\item{\code{output}}{The output generated when the model is run.
This is a list of dataframes.
\itemize{
  \item base:
    fundamental model output (see \link{CMetabDo})
  \item do:
    model output associated with dissolved oxygen calculations
    (see \link{CMetabDo})
  \item dic:
    model output associated with dissolved inorganic carbon calculations,
    a data frame with the following elements
    \itemize{
      \item pCO2:
        the partial pressure of dissolved carbon dioxide gas at
        the beginning of the time step (microatmospheres)
      \item dic:
        the concentration of dissolved inorganic carbon at the
        beginning of the time step (micromolarity)
      \item dicProduction:
        the number of dissolved inorganic carbon atoms generated by
        respiration during the time step per volume of water
        (micromolarity)
      \item dicConsumption:
        the number of dissolved inorganic carbon atoms consumed by
        primary production during the time step per volume of water
        (micromolarity)
      \item dicEquilibration:
        the net number of dissolved carbon dioxide gas molecules exchanged
        with the atomosphere during the time step per volume of water
        (micromolarity)
      \item kCO2:
        the gas exchange rate for carbon dioxide used during the time step
        (per day)
      \item pH:
        the negated log of hydrogen ion concentration
        (pH units)
      \item kH:
        Henry's constant for carbon dioxide used during the time step
        (molarity per atmosphere)
    }
}}

\item{\code{pCO2air}}{A vector of partial pressures of CO2 at the times when DO
values are to be simulated}

\item{\code{alkalinity}}{A vector of alkalinities due to carbonate at the times when DO
values are to be simulated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetabDoDic-new}{\code{CMetabDoDic$new()}}
\item \href{#method-CMetabDoDic-finalize}{\code{CMetabDoDic$finalize()}}
\item \href{#method-CMetabDoDic-run}{\code{CMetabDoDic$run()}}
\item \href{#method-CMetabDoDic-setMetabDoDicParam}{\code{CMetabDoDic$setMetabDoDicParam()}}
\item \href{#method-CMetabDoDic-clone}{\code{CMetabDoDic$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="metabc" data-topic="CMetab" data-id="setMetabParam"><a href='../../metabc/html/CMetab.html#method-CMetab-setMetabParam'><code>metabc::CMetab$setMetabParam()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabDoDic-new"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabDoDic-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$new(
  type = "ForwardEuler",
  dailyGPP,
  ratioDoCFix = 1,
  dailyER,
  ratioDoCResp = -1,
  k600,
  initialDO,
  time,
  temp,
  par,
  parTotal = -1,
  airPressure,
  stdAirPressure = 1,
  gwAlpha = NA,
  gwDO = NA,
  ratioDicCFix = -1,
  ratioDicCResp = 1,
  initialDIC,
  pCO2air,
  alkalinity,
  gwDIC = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "ForwardEuler" or "CrankNicolson" are known solutions)}

\item{\code{dailyGPP}}{Model parameter for total gross primary production
based on an effective concentration of DOC fixed.
Note the modifier "daily" is an artifact of models like this one
typically being used over a single photoperiod. The value for
total GPP provided via this argument is redistributed according to
the relative distribution of PAR over the analysis period.
Therefore, this number should reflect the total GPP over
however many photoperiods are being analyzed in a given simulation.
Units of micromoles per liter per day.}

\item{\code{ratioDoCFix}}{Ratio of DO molecules produced relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyER}}{Model parameter for daily the rate of aerobic ecosystem respiration
based on an effective concentration of DOC respired.
This is applied as a constant zero-order rate over the entire
period of analysis.
Units of micromoles per liter per day.}

\item{\code{ratioDoCResp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to -1.}

\item{\code{k600}}{Influence of atmospheric gas exchange on oxygen concentration
as a first-order rate depending on saturation deficit.
Units are per day.}

\item{\code{initialDO}}{Initial DO concentration at the beginning of the evaluation period.
Units are micromoles O2 per liter.}

\item{\code{time}}{Vector of times at which DO will be predicted.
These times must align with vectors provided for temperature and
photosynthetically active radiation.
Type should be a POSIXct vector or character vector that can be forced
to a POSIXct by default (e.g. format of "YYYY-MM-DD HH:mm:ss").}

\item{\code{temp}}{Vector of water temperatures in degrees Celsius.
Must be same length as the time vector.}

\item{\code{par}}{Vector of photosynthetically active radiation values.
Units are arbitrary if linear with PAR intensity, but must be consistent
with total PAR if the optional argument for total PAR is specified.
Must be the same length as the time vector.}

\item{\code{parTotal}}{Total PAR for period of analysis, see units note in par argument description.
If a negative value is provided, the model will integrate the par data
provided to estimate the total PAR.
Defaults to -1.}

\item{\code{airPressure}}{Barometric pressure.
If a single value is provided, a vector of the length of the time vector
will be created from that value.
Units must be consistent with the units of the optional standard air pressure
argument.
If default stdAirPressure is used, this should be a fraction of the standard
air pressure at sea level.}

\item{\code{stdAirPressure}}{The standard air pressure in the desired units.
Defaults to 1 standard atmosphere.}

\item{\code{gwAlpha}}{The turnover rate of channel water due to groundwater input.
Units of per day.
Default value is NA, which disables groundwater inflow simulation.
Can be a single value or a vector that provides a changing value over time.}

\item{\code{gwDO}}{The concentration of DO in inflowing groundwater.
Units of micromolarity.
Default value is NA, which disables groundwater inflow simulation.
Can be a single value or a vector that provides a changing value over time.}

\item{\code{ratioDicCFix}}{Ratio of DIC-C atoms consumed relative to organic carbon atoms fixed.
Defaults to -1.}

\item{\code{ratioDicCResp}}{Ratio of DIC-C atoms produced relative to organic carbon atoms respired
Defaults to 1.}

\item{\code{initialDIC}}{Initial DIC concentration at the beginning of the evaluation period.
Units are micromoles C per liter.}

\item{\code{pCO2air}}{A vector of partial pressures of CO2 at the times when DO
values are to be simulated.
If a single value is provided, a vector of the length of the time vector
will be created from that value.}

\item{\code{alkalinity}}{A vector of alkalinities due to carbonate at the times when DO
values are to be simulated.
If a single value is provided, a vector of the length of the time vector
will be created from that value.}

\item{\code{gwDIC}}{The concentration of DIC in inflowing groundwater.
Units of micromolarity of C.
Default value is NA, which disables groundwater inflow simulation.
Can be a single value or a vector that provides a changing value over time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabDoDic-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabDoDic-finalize}{}}}
\subsection{Method \code{finalize()}}{
Calls the R wrapper of the destructor of the
  underlying C object to release memory.
  Do not call this manually unless you really know what you are doing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SEXP object returned by the destructor method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabDoDic-run"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabDoDic-run}{}}}
\subsection{Method \code{run()}}{
Runs a simulation based on parameters and driving data
  in the object attributes. Output will be stored in the
  output attribute, but is also returned from this function
  as a convenience.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of dataframes containing model output. See description
  of the output attribute for details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabDoDic-setMetabDoDicParam"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabDoDic-setMetabDoDicParam}{}}}
\subsection{Method \code{setMetabDoDicParam()}}{
Sets one of the DO or DIC related model parameters to a new value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$setMetabDoDicParam(param, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A character string to indicate the parameter to be changed
\itemize{
  \item "RatioDoCFix": change the DO C fixation stoichiometric parameter
  \item "RatioDoCResp": change the DO C respiration stoichiometric parameter
  \item "RatioDicCFix": change the DIC C fixation stoichiometric parameter
  \item "RatioDicCResp": change the DIC C respiration stoichiometric parameter
}}

\item{\code{value}}{The new value for the parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The previous value of the parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabDoDic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabDoDic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
