% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabExtractor.R
\name{CMetabExtractor}
\alias{CMetabExtractor}
\title{R6 class for extracting results from one station metabolism models}
\description{
Extracts parameters used from metabolism models
}
\section{Super class}{
\code{\link[disco:SignalSummarizer]{disco::SignalSummarizer}} -> \code{CMetabExtractor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{table}}{Table of summary data}

\item{\code{index}}{Index for tracking the row in the table}

\item{\code{resultFile}}{The name of the file with analysis results}

\item{\code{summaryFile}}{The name of the file to create with the summary}

\item{\code{path}}{Table of summary data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetabExtractor-new}{\code{CMetabExtractor$new()}}
\item \href{#method-CMetabExtractor-open}{\code{CMetabExtractor$open()}}
\item \href{#method-CMetabExtractor-summarize}{\code{CMetabExtractor$summarize()}}
\item \href{#method-CMetabExtractor-close}{\code{CMetabExtractor$close()}}
\item \href{#method-CMetabExtractor-clone}{\code{CMetabExtractor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabExtractor-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabExtractor$new(rows, resultFile = "results", summaryFile = "summary")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{number of rows in the table of extracted parameter values}

\item{\code{resultFile}}{name of the results files to read}

\item{\code{summaryFile}}{name of the summary file to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabExtractor-open"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabExtractor-open}{}}}
\subsection{Method \code{open()}}{
Prepare to write the summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabExtractor$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path where the summary file will be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabExtractor-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabExtractor-summarize}{}}}
\subsection{Method \code{summarize()}}{
Collects the data for the summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabExtractor$summarize(signal, outputPath, label, timeBounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{The signal being summarized}

\item{\code{outputPath}}{The path where analysis output is located}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabExtractor-close"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabExtractor-close}{}}}
\subsection{Method \code{close()}}{
Closes the summary output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabExtractor$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
