% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabLagrangeDo.R
\name{CMetabLagrangeDo}
\alias{CMetabLagrangeDo}
\title{Lagrangian parcel DO metabolism model based on C++ implementation}
\description{
Predicts the change in dissolved oxygen
  in a parcel of water passing through an aquatic ecosystem due to
  metabolic processes and gas exchange with the atmosphere.

  Implementation of this model is provided by C++ objects.
}
\section{Super class}{
\code{\link[metabc:CMetab]{metabc::CMetab}} -> \code{CMetabLagrangeDo}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "CNOneStep" is a known solution using a single
calculation with Crank Nicolson type approximations)}

\item{\code{upstreamTimePOSIX}}{A vector of POSIX objects representing the times the
parcels passed the upstream end of the system}

\item{\code{downstreamTimePOSIX}}{A vector of POSIX objects representing the times the
parcels passed the downstream end of the system}

\item{\code{output}}{The output generated when the model is run.
This is a list of dataframes with the following named elements.
\itemize{
  \item base:
    fundamental model output data frame with the following columns
    \itemize{
      \item travelTimes:
        the time of travel for each water parcel (days)
      \item cFixation:
        the number of atoms of carbon fixed during the
        travel time (micromoles per liter)
      \item cRespiration:
        the number of atoms of carbon respired during the
        travel time (micromoles per liter)
    }
  \item do:
    model output associated with dissolved oxygen calculations,
    a data frame with the following elements
    \itemize{
      \item dox:
        the concentration of dioxygen gas as the parcel passes the
        downstream location (micromolarity)
      \item upstreamDoSat:
        the saturation concentration of dioxygen gas as the parcel
        passes the upstream location (micromolarity)
      \item downstreamDoSat:
        the saturation concentration of dioxygen gas as the parcel
        passes the downstream location (micromolarity)
      \item doProduction:
        the number of dioxygen gas molecules generated by
        primary production during the travel time per volume of water
        (micromolarity)
      \item doConsumption:
        the number of dioxygen gas molecules consumed by
        respiration during the travel time per volume of water
        (micromolarity)
      \item upstreamkDo:
        the gas exchange rate for oxygen as the parcel passes the upstream
        location
        (per day)
      \item downstreamkDo:
        the gas exchange rate for oxygen as the parcel passes the downstream
        location
        (per day)
      \item doEquilibration:
        the numerator term associated with the net number of dissolved carbon
        dioxide gas molecules exchanged
        with the atomosphere during the travel time per volume of water
        (micromolarity)
    }
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetabLagrangeDo-new}{\code{CMetabLagrangeDo$new()}}
\item \href{#method-CMetabLagrangeDo-finalize}{\code{CMetabLagrangeDo$finalize()}}
\item \href{#method-CMetabLagrangeDo-run}{\code{CMetabLagrangeDo$run()}}
\item \href{#method-CMetabLagrangeDo-setMetabLagrangeDoParam}{\code{CMetabLagrangeDo$setMetabLagrangeDoParam()}}
\item \href{#method-CMetabLagrangeDo-clone}{\code{CMetabLagrangeDo$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="metabc" data-topic="CMetab" data-id="setMetabParam"><a href='../../metabc/html/CMetab.html#method-CMetab-setMetabParam'><code>metabc::CMetab$setMetabParam()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeDo-new"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeDo-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeDo$new(
  type = "CNOneStep",
  dailyGPP,
  ratioDoCFix = 1,
  dailyER,
  ratioDoCResp = -1,
  k600,
  upstreamDO,
  upstreamTime,
  downstreamTime,
  upstreamTemp,
  downstreamTemp,
  upstreamPAR,
  downstreamPAR,
  parTotal = -1,
  airPressure,
  stdAirPressure = 1,
  timesteps = 2,
  gwAlpha = NA,
  gwDO = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "CNOneStep" is a known solution using a single
calculation with Crank Nicolson type approximations)}

\item{\code{dailyGPP}}{Model parameter for daily gross primary production
based on an effective concentration of DOC fixed.
Units of micromoles per liter per day.}

\item{\code{ratioDoCFix}}{Ratio of DO molecules produced relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyER}}{Model parameter for daily aerobic ecosystem respiration
based on an effective concentration of DOC respired
Units of micromoles per liter per day.}

\item{\code{ratioDoCResp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to -1.}

\item{\code{k600}}{Influence of atmospheric gas exchange on oxygen concentration
as a first-order rate depending on saturation deficit.
Units are per day.}

\item{\code{upstreamDO}}{Vector of DO concentrations when the parcels of water are passing
the upstream end of the system}

\item{\code{upstreamTime}}{Vector of times when parcels are passing upstream
These times must align with vectors provided for temperature and
photosynthetically active radiation.
Type should be a POSIXct vector or character vector that can be forced
to a POSIXct by default (e.g. format of "YYYY-MM-DD HH:mm:ss").}

\item{\code{downstreamTime}}{Vector of times when parcels are passing downstream
These times must align with vectors provided for temperature and
photosynthetically active radiation.
Type should be a POSIXct vector or character vector that can be forced
to a POSIXct by default (e.g. format of "YYYY-MM-DD HH:mm:ss").}

\item{\code{upstreamTemp}}{Vector of water temperatures in parcels passing upstream.
Must be same length as the time vectors.}

\item{\code{downstreamTemp}}{Vector of water temperatures in parcels passing downstream.
Must be same length as the time vectors.}

\item{\code{upstreamPAR}}{Vector of photosynthetically active radiation values on parcels
passing upstream.
Units are arbitrary if linear with PAR intensity, but must be consistent
with total PAR if the optional argument for total PAR is specified.
Must be the same length as the time vector.}

\item{\code{downstreamPAR}}{Vector of photosynthetically active radiation values on parcels
passing downstream.
Units are arbitrary if linear with PAR intensity, but must be consistent
with total PAR if the optional argument for total PAR is specified.
Must be the same length as the time vector.}

\item{\code{parTotal}}{Total PAR for period of analysis, see units note in par argument description.
If a negative value is provided, the model will integrate the par data
provided to estimate the total PAR.
Defaults to -1.}

\item{\code{airPressure}}{Barometric pressure. A vector for the parcels can be provided, or a
single value will be automatically turned into a
vector of values for the parcels.
Units must be consistent with the units of the optional standard air pressure
argument.
If default stdAirPressure is used, this should be a fraction of the standard
air pressure at sea level.}

\item{\code{stdAirPressure}}{The standard air pressure in the desired units.
Defaults to 1 standard atmosphere.}

\item{\code{timesteps}}{The number of timesteps for calculations for each parcel.
Note that certain calculation algorithms (e.g. "OneStep" calculators)
will ignore this argument.
Defaults to 2 (a single calculation per parcel).}

\item{\code{gwAlpha}}{The turnover rate of channel water due to groundwater input.
Units of per day.
Default value is NA, which disables groundwater inflow simulation.
Can be a single value or a vector that provides a changing value over time.}

\item{\code{gwDO}}{The concentration of DO in inflowing groundwater.
Units of micromolarity.
Default value is NA, which disables groundwater inflow simulation.
Can be a single value or a vector that provides a changing value over time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeDo-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeDo-finalize}{}}}
\subsection{Method \code{finalize()}}{
Calls the R wrapper of the destructor of the
  underlying C object to release memory.
  Do not call this manually unless you really know what you are doing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeDo$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SEXP object returned by the destructor method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeDo-run"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeDo-run}{}}}
\subsection{Method \code{run()}}{
Runs a simulation based on parameters and driving data
  in the object attributes. Output will be stored in the
  output attribute, but is also returned from this function
  as a convenience.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeDo$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of dataframes containing model output.
  See description of output attribute for details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeDo-setMetabLagrangeDoParam"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeDo-setMetabLagrangeDoParam}{}}}
\subsection{Method \code{setMetabLagrangeDoParam()}}{
Sets one of the DO related model parameters to a new value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeDo$setMetabLagrangeDoParam(param, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A character string to indicate the parameter to be changed
\itemize{
  \item "RatioDoCFix": change the C fixation stoichiometric parameter
  \item "RatioDoCResp": change the C respiration stoichiometric parameter
}}

\item{\code{value}}{The new value for the parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The previous value of the parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeDo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeDo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeDo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
