% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabLagrangeExtractor.R
\name{CMetabLagrangeExtractor}
\alias{CMetabLagrangeExtractor}
\title{R6 class for extracting results from one station metabolism models}
\description{
Extracts parameters used from metabolism models
}
\section{Super class}{
\code{\link[disco:TransferFunctionSummarizer]{disco::TransferFunctionSummarizer}} -> \code{CMetabLagrangeExtractor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{table}}{Table of summary data}

\item{\code{index}}{Index for tracking the row in the table}

\item{\code{resultFile}}{The name of the file with analysis results}

\item{\code{summaryFile}}{The name of the file to create with the summary}

\item{\code{path}}{Table of summary data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetabLagrangeExtractor-new}{\code{CMetabLagrangeExtractor$new()}}
\item \href{#method-CMetabLagrangeExtractor-open}{\code{CMetabLagrangeExtractor$open()}}
\item \href{#method-CMetabLagrangeExtractor-summarize}{\code{CMetabLagrangeExtractor$summarize()}}
\item \href{#method-CMetabLagrangeExtractor-close}{\code{CMetabLagrangeExtractor$close()}}
\item \href{#method-CMetabLagrangeExtractor-clone}{\code{CMetabLagrangeExtractor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeExtractor-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeExtractor$new(
  rows,
  resultFile = "results",
  summaryFile = "summary"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{number of rows in the table of extracted parameter values}

\item{\code{resultFile}}{name of the results files to read}

\item{\code{summaryFile}}{name of the summary file to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeExtractor-open"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeExtractor-open}{}}}
\subsection{Method \code{open()}}{
Prepare to write the summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeExtractor$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path where the summary file will be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeExtractor-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeExtractor-summarize}{}}}
\subsection{Method \code{summarize()}}{
Collects the data for the summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeExtractor$summarize(
  signalIn,
  signalOut,
  outputPath,
  label,
  timeBounds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{The input signal for the transfer function being summarized}

\item{\code{signalOut}}{The output signal for the transfer function being summarized}

\item{\code{outputPath}}{The path where analysis output is located}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeExtractor-close"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeExtractor-close}{}}}
\subsection{Method \code{close()}}{
Closes the summary output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeExtractor$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabLagrangeExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabLagrangeExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabLagrangeExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
