% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppOneStationMetabDoDic.R
\name{CppOneStationMetabDoDic}
\alias{CppOneStationMetabDoDic}
\title{Construct a one-station DO model object}
\value{
Pointer to an instance of a one-station DO model object
}
\description{
Construct a one-station DO model object

Construct a one-station DO model object
}
\section{Super classes}{
\code{\link[metab:OneStationMetabDo]{metab::OneStationMetabDo}} -> \code{\link[metab:OneStationMetabDoDic]{metab::OneStationMetabDoDic}} -> \code{CppOneStationMetabDoDic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CppOneStationMetabDoDic$new()}}
\item \href{#method-finalize}{\code{CppOneStationMetabDoDic$finalize()}}
\item \href{#method-run}{\code{CppOneStationMetabDoDic$run()}}
\item \href{#method-clone}{\code{CppOneStationMetabDoDic$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="getDailyErDo">}\href{../../metab/html/OneStationMetabDo.html#method-getDailyErDo}{\code{metab::OneStationMetabDo$getDailyErDo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="getDailyGppDo">}\href{../../metab/html/OneStationMetabDo.html#method-getDailyGppDo}{\code{metab::OneStationMetabDo$getDailyGppDo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="plot">}\href{../../metab/html/OneStationMetabDo.html#method-plot}{\code{metab::OneStationMetabDo$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="plotPAR">}\href{../../metab/html/OneStationMetabDo.html#method-plotPAR}{\code{metab::OneStationMetabDo$plotPAR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDoDic" data-id="getDailyErDic">}\href{../../metab/html/OneStationMetabDoDic.html#method-getDailyErDic}{\code{metab::OneStationMetabDoDic$getDailyErDic()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDoDic" data-id="getDailyGppDic">}\href{../../metab/html/OneStationMetabDoDic.html#method-getDailyGppDic}{\code{metab::OneStationMetabDoDic$getDailyGppDic()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppOneStationMetabDoDic$new(
  ...,
  time,
  airPressure,
  stdAirPressure = 760,
  pCO2air,
  alkalinity
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppOneStationMetabDoDic$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppOneStationMetabDoDic$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CppOneStationMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
