% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfernoParameterTranslator.R
\name{ParameterTranslatorMetabc}
\alias{ParameterTranslatorMetabc}
\title{R6 class defining a parameter translator for metabolism models}
\description{
A parameter translator allowing GPP, ER, and k600 to be used
   as proposed parameters for an objective function using a
   stream metabolism model.
}
\section{Super class}{
\code{\link[inferno:ParameterTranslator]{inferno::ParameterTranslator}} -> \code{ParameterTranslatorMetabc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-translate}{\code{ParameterTranslatorMetabc$translate()}}
\item \href{#method-clone}{\code{ParameterTranslatorMetabc$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ParameterTranslator" data-id="initialize">}\href{../../inferno/html/ParameterTranslator.html#method-initialize}{\code{inferno::ParameterTranslator$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\if{latex}{\out{\hypertarget{method-translate}{}}}
\subsection{Method \code{translate()}}{
Used to translate 3 parameters in the order of
  gross primary production, ecosystem respiration,
  and gas exchange at a Schmidt number of 600
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterTranslatorMetabc$translate(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Named vector of values to use for GPP (named "gpp"),
ER (named "er"), and k600 (name "er") parameters.
Parameters with missing names will not be translated.
Any additional elements will be ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterTranslatorMetabc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
