% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfernoPredictionExtractor.R
\name{PredictionExtractorMetabc}
\alias{PredictionExtractorMetabc}
\title{R6 class defining a prediction extractor for DO from metabolism models}
\description{
A prediction extractor allowing dissolved oxygen outputs
   from a stream metabolism model to be used as predictions
   in an objective function.
}
\section{Super class}{
\code{\link[inferno:PredictionExtractor]{inferno::PredictionExtractor}} -> \code{PredictionExtractorMetabc}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{extractDO}}{Logical value indicating if DO predictions should be extracted}

\item{\code{extractCO2}}{Logical value indicating if pCO2 predictions should be extracted}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictionExtractorMetabc-new}{\code{PredictionExtractorMetabc$new()}}
\item \href{#method-PredictionExtractorMetabc-extract}{\code{PredictionExtractorMetabc$extract()}}
\item \href{#method-PredictionExtractorMetabc-clone}{\code{PredictionExtractorMetabc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="inferno" data-topic="PredictionExtractor" data-id="setModel"><a href='../../inferno/html/PredictionExtractor.html#method-PredictionExtractor-setModel'><code>inferno::PredictionExtractor$setModel()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExtractorMetabc-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExtractorMetabc-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the class with provided values for
  extraction flags
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractorMetabc$new(..., extractDO, extractCO2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to PredictionExtractor super class}

\item{\code{extractDO}}{Logical value indicating if DO predictions should be extracted}

\item{\code{extractCO2}}{Logical value indicating if pCO2 predictions should be extracted}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExtractorMetabc-extract"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExtractorMetabc-extract}{}}}
\subsection{Method \code{extract()}}{
Used to extract dissolved oxygen predictions from a metabolism model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractorMetabc$extract()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with a single column of DO data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictionExtractorMetabc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictionExtractorMetabc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractorMetabc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
