% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabDoDic.R
\name{CMetabDoDic}
\alias{CMetabDoDic}
\title{Time series DO and DIC metabolism model based on C++ implementation}
\description{
Predicts the change in dissolved oxygen and dissolved inorganic carbon
  in an aquatic system over time due to metabolic processes and
  gas exchange with the atmosphere.

  Implementation of this model is provided by C++ objects.
}
\section{Super class}{
\code{\link[metabc:CMetab]{metabc::CMetab}} -> \code{CMetabDoDic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "ForwardEuler" or "CrankNicolson" are known solutions)}

\item{\code{time}}{A numeric vector of time in units of days representing the
times at which DO concentrations are simulated}

\item{\code{timePOSIX}}{A vector of POSIXct type time objects representing the
times at which DO concentrations are simulated}

\item{\code{airPressure}}{A vector of barometric pressures at the times when DO
values are to be simulated}

\item{\code{output}}{The output generated when the model is run.
This is a list of dataframes.}

\item{\code{pCO2air}}{A vector of partial pressures of CO2 at the times when DO
values are to be simulated}

\item{\code{alkalinity}}{A vector of alkalinities due to carbonate at the times when DO
values are to be simulated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CMetabDoDic$new()}}
\item \href{#method-finalize}{\code{CMetabDoDic$finalize()}}
\item \href{#method-run}{\code{CMetabDoDic$run()}}
\item \href{#method-setMetabDoDicParam}{\code{CMetabDoDic$setMetabDoDicParam()}}
\item \href{#method-clone}{\code{CMetabDoDic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="metabc" data-topic="CMetab" data-id="setMetabParam">}\href{../../metabc/html/CMetab.html#method-setMetabParam}{\code{metabc::CMetab$setMetabParam()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$new(
  type = "ForwardEuler",
  dailyGPP,
  ratioDoCFix = 1,
  dailyER,
  ratioDoCResp = -1,
  k600,
  initialDO,
  time,
  temp,
  par,
  parTotal = -1,
  airPressure,
  stdAirPressure = 1,
  ratioDicCFix = -1,
  ratioDicCResp = 1,
  initialDIC,
  pCO2air,
  alkalinity
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A character string indicating type of solution used.
(e.g. "ForwardEuler" or "CrankNicolson" are known solutions)}

\item{\code{dailyGPP}}{Model parameter for daily gross primary production
based on an effective concentration of DOC fixed.
Units of micromoles per liter per day.}

\item{\code{ratioDoCFix}}{Ratio of DO molecules produced relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyER}}{Model parameter for daily aerobic ecosystem respiration
based on an effective concentration of DOC respired
Units of micromoles per liter per day.}

\item{\code{ratioDoCResp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to -1.}

\item{\code{k600}}{Influence of atmospheric gas exchange on oxygen concentration
as a first-order rate depending on saturation deficit.
Units are per day.}

\item{\code{initialDO}}{Initial DO concentration at the beginning of the evaluation period.
Units are micromoles O2 per liter.}

\item{\code{time}}{Vector of times at which DO will be predicted.
These times must align with vectors provided for temperature and
photosynthetically active radiation.
Type should be a POSIXct vector or character vector that can be forced
to a POSIXct by default (e.g. format of "YYYY-MM-DD HH:mm:ss").}

\item{\code{temp}}{Vector of water temperatures in degrees Celsius.
Must be same length as the time vector.}

\item{\code{par}}{Vector of photosynthetically active radiation values.
Units are arbitrary if linear with PAR intensity, but must be consistent
with total PAR if the optional argument for total PAR is specified.
Must be the same length as the time vector.}

\item{\code{parTotal}}{Total PAR for period of analysis, see units note in par argument description.
If a negative value is provided, the model will integrate the par data
provided to estimate the total PAR.
Defaults to -1.}

\item{\code{airPressure}}{Barometric pressure.
If a single value is provided, a vector of the length of the time vector
will be created from that value.
Units must be consistent with the units of the optional standard air pressure
argument.
If default stdAirPressure is used, this should be a fraction of the standard
air pressure at sea level.}

\item{\code{stdAirPressure}}{The standard air pressure in the desired units.
Defaults to 1 standard atmosphere.}

\item{\code{ratioDicCFix}}{Ratio of DIC-C atoms consumed relative to organic carbon atoms fixed.
Defaults to -1.}

\item{\code{ratioDicCResp}}{Ratio of DIC-C atoms produced relative to organic carbon atoms respired
Defaults to 1.}

\item{\code{initialDIC}}{Initial DIC concentration at the beginning of the evaluation period.
Units are micromoles C per liter.}

\item{\code{pCO2air}}{A vector of partial pressures of CO2 at the times when DO
values are to be simulated.
If a single value is provided, a vector of the length of the time vector
will be created from that value.}

\item{\code{alkalinity}}{A vector of alkalinities due to carbonate at the times when DO
values are to be simulated.
If a single value is provided, a vector of the length of the time vector
will be created from that value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Calls the R wrapper of the destructor of the
  underlying C object to release memory.
  Do not call this manually unless you really know what you are doing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SEXP object returned by the destructor method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Runs a simulation based on parameters and driving data
  in the object attributes. Output will be stored in the
  output attribute, but is also returned from this function
  as a convenience.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of dataframes containing model output.
  \itemize{
    \item base: fundamental model output
    \item do: model output associated with dissolved oxygen calculations
    \item dic: model output associated with dissolved inorganic carbon calculations
  }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMetabDoDicParam"></a>}}
\if{latex}{\out{\hypertarget{method-setMetabDoDicParam}{}}}
\subsection{Method \code{setMetabDoDicParam()}}{
Sets one of the DO or DIC related model parameters to a new value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$setMetabDoDicParam(param, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A character string to indicate the parameter to be changed
\itemize{
  \item "RatioDoCFix": change the DO C fixation stoichiometric parameter
  \item "RatioDoCResp": change the DO C respiration stoichiometric parameter
  \item "RatioDicCFix": change the DIC C fixation stoichiometric parameter
  \item "RatioDicCResp": change the DIC C respiration stoichiometric parameter
}}

\item{\code{value}}{The new value for the parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The previous value of the parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
