% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMetabOptim.R
\name{CMetabOptim}
\alias{CMetabOptim}
\title{R6 class defining an MLE inference using a metabolism model at a single location}
\description{
Processes a signal of DO and pCO2 data
  (with supporting PAR and temperature information)
  for inferring whole-stream metabolism using the one-station
  method.
}
\section{Super class}{
\code{\link[disco:SignalDerivation]{disco::SignalDerivation}} -> \code{CMetabOptim}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initParams}}{The intial parameter values to use for the MLE algorithm}

\item{\code{fixedParams}}{A named list of values for non-estimated parameters}

\item{\code{objFunc}}{The objective function to use with optim for the inference}

\item{\code{modelType}}{Character string representing the type of model calculation to use}

\item{\code{staticAirPressure}}{The air pressure.}

\item{\code{staticGwAlpha}}{Static value for the channel turnover rate due to inflowing groundwater}

\item{\code{staticGwDO}}{Static value for DO concentration in inflowing groundwater}

\item{\code{useDO}}{Logical flag to indicate whether DO data should be used for inference}

\item{\code{doHeader}}{character string representing the header for DO to use. Null
value will prevent DO from being used in analysis.}

\item{\code{usepCO2}}{Logical flag to indicate whether pCO2 data should be used for inference}

\item{\code{pCO2Header}}{Character string representing the header for pCO2 to use. Null
value will prevent pCO2 from being used in analysis.}

\item{\code{dicHeader}}{Character string representing the header for DIC to use.}

\item{\code{parHeader}}{Character string representing the header for PAR to use.}

\item{\code{tempHeader}}{Character string representing the header for temperature to use.}

\item{\code{airPressureHeader}}{Character string representing the header for air pressure to use.}

\item{\code{gwAlphaHeader}}{Character string representing the header for
channel turnover rate due to inflowing groundwater.}

\item{\code{gwDOHeader}}{Character string representing the header for DO in inflowing groundwater}

\item{\code{outputFile}}{The name of the output file.}

\item{\code{staticCO2Air}}{The pCO2 in the air.}

\item{\code{staticAlkalinity}}{The alkalinity.}

\item{\code{staticGwDIC}}{Static value for DIC in groundwater}

\item{\code{co2AirHeader}}{Character string representing the header for co2 partial pressure in air to use.}

\item{\code{alkalinityHeader}}{Character string representing the header for alkalinity to use.}

\item{\code{gwDICHeader}}{Character string representing the header for DIC in groundwater}

\item{\code{optimArgs}}{A list representing arguments to pass to optim}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CMetabOptim-new}{\code{CMetabOptim$new()}}
\item \href{#method-CMetabOptim-derive}{\code{CMetabOptim$derive()}}
\item \href{#method-CMetabOptim-clone}{\code{CMetabOptim$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabOptim-new"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabOptim-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new object of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabOptim$new(
  ...,
  initParams,
  fixedParams = NULL,
  objFunc,
  modelType = "ForwardEuler",
  useDO,
  doHeader = "do",
  usepCO2,
  pCO2Header = "pCO2",
  dicHeader = "dic",
  parHeader = "par",
  tempHeader = "temp",
  staticAirPressure = NULL,
  airPressureHeader = "airPressure",
  outputFile = "results",
  staticCO2Air = NULL,
  co2AirHeader = "co2Air",
  staticAlkalinity = NULL,
  alkalinityHeader = "alkalinity",
  staticGwAlpha = NULL,
  gwAlphaHeader = NULL,
  staticGwDO = NULL,
  gwDOHeader = NULL,
  staticGwDIC = NULL,
  gwDICHeader = NULL,
  optimArgs = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to the initialize method of the super class.
See \link{CMetabOptim} for documentation on these parameters}

\item{\code{initParams}}{The initial parameter values to use in the MLE algorithm.
A named vector of numeric values should be provided.
Parameter names currently supported include
dailyGPP, dailyER, and k600.}

\item{\code{fixedParams}}{An optional named list of values for non-estimated parameters.
A named vector of numeric values should be provided.
Parameter names currently supported include
dailyGPP, dailyER, and k600.
Defaults to NULL, which indicates no fixed parameters.}

\item{\code{objFunc}}{The objective function to use with optim for the inference}

\item{\code{modelType}}{Character string representing the type of model calculation to use}

\item{\code{useDO}}{Logical flag to indicate whether DO data should be used for inference}

\item{\code{doHeader}}{character string representing the header for DO to use. Null
value will prevent DO from being used in analysis.}

\item{\code{usepCO2}}{Logical flag to indicate whether pCO2 data should be used for inference}

\item{\code{pCO2Header}}{Character string representing the header for pCO2 to use. Null
value will prevent pCO2 from being used in analysis.}

\item{\code{dicHeader}}{Character string representing the header for DIC to use.}

\item{\code{parHeader}}{Character string representing the header for PAR to use.}

\item{\code{tempHeader}}{Character string representing the header for temperature to use.}

\item{\code{staticAirPressure}}{The air pressure.}

\item{\code{airPressureHeader}}{Character string representing the header for air pressure to use.}

\item{\code{outputFile}}{The name of the output file.}

\item{\code{staticCO2Air}}{The pCO2 in the air.}

\item{\code{co2AirHeader}}{Character string representing the header for co2 partial pressure in air to use.}

\item{\code{staticAlkalinity}}{The alkalinity.}

\item{\code{alkalinityHeader}}{Character string representing the header for alkalinity to use.}

\item{\code{staticGwAlpha}}{Static value for the channel turnover rate due to inflowing groundwater.
Default value is NULL, meaning values will be taken from the
variable specified by gwAlphaHeader in the signal.}

\item{\code{gwAlphaHeader}}{Character string representing the header for
channel turnover rate due to inflowing groundwater.
Default value is NULL.
When both static value and header are NULL, simulation of
groundwater influence is disabled.}

\item{\code{staticGwDO}}{Static value for DO concentration in inflowing groundwater.
Default value is NULL, meaning values will be taken from the
variable specified by gwDOHeader in the signal.}

\item{\code{gwDOHeader}}{Character string representing the header for DO in inflowing groundwater
Default value is NULL.
When both static value and header are NULL, simulation of
groundwater influence is disabled.}

\item{\code{staticGwDIC}}{Static value for DIC in groundwater.
Default value is NULL, meaning values will be taken from the
variable specified by gwDICHeader in the signal.}

\item{\code{gwDICHeader}}{Character string representing the header for DIC in groundwater
Default value is NULL.
When both static value and header are NULL, simulation of
groundwater influence is disabled.}

\item{\code{optimArgs}}{A list representing arguments to pass to optim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabOptim-derive"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabOptim-derive}{}}}
\subsection{Method \code{derive()}}{
Performs a metabolism analysis using the minimization of a
  value from a provided objective function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabOptim$derive(signal = NULL, prevResults = NULL, path, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{The signal with the DO or pCO2 on which the metabolism estimate is based.
If null, the signal attribute will be used.}

\item{\code{prevResults}}{If available, the results from the previous window of analysis.}

\item{\code{path}}{The path to the location where the output file is written}

\item{\code{index}}{The index of the row of the data frame to use for any fixed
parameter values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The results object from the analysis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CMetabOptim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CMetabOptim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CMetabOptim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
